Fixed issues with dropdown resetting when tool bar closed and toast errors
The dropdown needed extra values sent down to have it work correctly. I added an extra state in the store "surahId" that storesthe index of the surah selected. I made changes to the translation input to mak sure it doesn't reset when the toolbar is closed

There was a bug in the toast. The ayaNumber state's start and end were beiong set as strings not numerics. When doing teh comparision to see if the toast should be sent string comparssion was behaving differently. I forced the state to be set as ints